agents/workshop_agent/requirements.txt
google-cloud-aiplatform[adk,agent_engines]==1.91.0
google-adk==0.4.0
mcp[cli]==1.7.1
aiohttp==3.9.5